requirements.txt
# SynthSearch-Biomed Requirements
# Python 3.10-3.12

# Core
numpy>=1.24,<2.0
pandas>=2.0
pyyaml>=6.0
tqdm>=4.65
requests>=2.31
aiohttp>=3.9

# Data processing
lxml>=4.9              # XML parsing for PMC
beautifulsoup4>=4.12   # HTML/XML parsing
jsonlines>=4.0         # JSONL I/O

# Retrieval - BM25
pyserini>=0.22         # BM25 via Lucene (requires Java 11+)
# rank-bm25>=0.2.2     # lightweight alternative (uncomment if no Java)

# Retrieval - Dense
transformers>=4.40
torch>=2.1
sentence-transformers>=3.0  # for SPECTER2 and reranker
faiss-cpu>=1.7.4       # vector index (use faiss-gpu if GPU available)

# Evaluation
scikit-learn>=1.3
scipy>=1.11

# Visualization
matplotlib>=3.8
seaborn>=0.13

# APIs
biopython>=1.83        # PubMed E-utilities (Bio.Entrez)

# Utilities
python-dotenv>=1.0
loguru>=0.7
